requirements.txt
requests==2.31.0
newsapi-python==0.2.6
python-dotenv==1.0.0
